requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cryptography==35.0.0
cycler==0.10.0
DAWG-Python==0.7.2
docopt==0.6.2
idna==3.2
intervaltree==3.1.0
ipymarkup==0.9.0
kiwisolver==1.3.2
matplotlib==3.4.3
natasha==1.4.0
navec==0.10.0
numpy==1.21.2
pdfminer.six==20201018
Pillow==8.3.2
pycparser==2.20
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
python-dateutil==2.8.2
razdel==0.5.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
slovnet==0.5.0
sortedcontainers==2.4.0
soupsieve==2.2.1
urllib3==1.26.7
wordcloud==1.8.1
yargy==0.15.0
